feat: add "at" keyword for imshow and other enhancements (#1430)
* guess plot locations if not provided through 'at' keyword

* add unit tests for location guessing

* use 'at' keyword for RasterModelGrid.imshow

* move imshowhs functions into their own module

* import RasterModelGrid within query_grid_on_button_press to avoid circular import

* rename RasterModelGridPlotter to ModelGridPlotterMixIn

* import RasterModelGrid within _imshowhs_grid_values to avoid circular import

* update some notebooks and documentation for imshow

* add missing import of RasterModelGrid

* add news fragments

[skip ci]
