+++ news/1424.feature
@@ -1,0 +1,3 @@
+Added the ``at`` keyword to the ``imshow_grid`` functions so that they now
+use the same pattern as many other *landlab* functions.
+
