+++ news/1430.feature
@@ -1,0 +1,3 @@
+Added an ``imshow`` method to all *landlab* model grids that is a wrapper for
+the ``imshow_grid`` function.
+
